+++ assets/report-system.js
@@ -1,7 +1,7 @@
 (() => {
   "use strict";
 
-  const VERSION = "1.2.33";
+  const VERSION = "1.2.34";
 
   function escapeHtml(value) {
     return String(value ?? "").replace(/[&<>'"]/g, char => ({ "&":"&amp;", "<":"&lt;", ">":"&gt;", "'":"&#39;", '"':"&quot;" }[char]));